requirement.txt
Flask==3.1.0
mistralai==1.5.0
python-dotenv==1.0.1